requirements.txt
annotated-types
asgiref
cachetools
certifi
cffi
charset-normalizer
cryptography
Django
djangorestframework
docstring_parser
drf-yasg
google-api-core
google-auth
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-core
google-cloud-resource-manager
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
grpc-google-iam-v1
grpcio
grpcio-status
idna
inflection
numpy
pymssql
packaging
pandas
proto-plus
protobuf
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
python-dateutil
python-dotenv
pytz
PyYAML
requests
rsa
shapely
six
sqlparse
typing_extensions
tzdata
uritemplate
urllib3
vertexai
django-csp